backend/cart-service/requirements.txt
Django>=4.2,<5.0
djangorestframework
redis
